Move implementations from __init__
